requirements.txt
pandas>=2.0
numpy>=1.25
matplotlib>=3.8
scikit-learn>=1.3
nltk>=3.8.1
scikit-fuzzy>=0.4.2
pyyaml>=6.0
tqdm>=4.66
